Also move ssl_context and while there rename
