+++ src/libmeasurement_kit/libevent/ssl_context.cpp
@@ -9,10 +9,10 @@
 #include <stdexcept>
 #include <iostream>
 #include <measurement_kit/net.hpp>
-#include "../net/ssl-context.hpp"
+#include "../libevent/ssl_context.hpp"
 
 namespace mk {
-namespace net {
+namespace libevent {
 
 ErrorOr<SSL *> SslContext::get_client_ssl(std::string hostname) {
     SSL *ssl = SSL_new(ctx);
@@ -79,6 +79,6 @@ SslContext::~SslContext() {
     SSL_CTX_free(ctx);
 }
 
-} // namespace net
+} // namespace libevent
 } // namespace mk
